+++ libs/cart/routing/src/guards/order-result/order-result-guard-redirect.token.ts
