+++ libs/cart/state/src/public_api.ts
@@ -1,14 +1,10 @@
 export * from './actions/public_api';
 export * from './selectors/public_api';
 export * from './reducers/public_api';
-export * from './guards/public_api';
-export * from './resolvers/public_api';
 export * from './models/public_api';
 export * from './injection-tokens/public_api';
 
 export { DaffCartStateModule } from './cart-state.module';
-export { DaffCartStateConfiguration } from './config/config';
-export { DAFF_CART_STATE_CONFIG } from './config/config';
 
 export { DaffCartFacade } from './facades/cart/cart.facade';
 export { DaffCartFacadeInterface } from './facades/cart/cart-facade.interface';
